Remove spawn zones and refactor spawn positioning to use ship boundary
