+++ src/AppOne.ts
@@ -470,15 +470,6 @@ export class AppOne {
                 )
             })
 
-        const spawnZonesState = { enabled: false }
-        const spawnZonesController = debugFolder
-            .add(spawnZonesState, 'enabled')
-            .name('Show Spawn Zones')
-            .onChange((enabled: boolean) => {
-                this.gameWorld.toggleDebugSpawnZones(enabled)
-                console.log(`🟢 Spawn Zones: ${enabled ? 'Visible' : 'Hidden'}`)
-            })
-
         const boundariesState = { enabled: false }
         const boundariesController = debugFolder
             .add(boundariesState, 'enabled')
@@ -499,7 +490,6 @@ export class AppOne {
                         weaponRangeState.enabled = true
                         velocityVectorsState.enabled = true
                         restrictedZonesState.enabled = true
-                        spawnZonesState.enabled = true
                         boundariesState.enabled = true
 
                         this.gameWorld.setDebugMode(true)
@@ -508,7 +498,6 @@ export class AppOne {
                         this.gameWorld.toggleDebugWeaponRange(true)
                         this.gameWorld.toggleDebugVelocityVectors(true)
                         this.gameWorld.toggleDebugRestrictedZones(true)
-                        this.gameWorld.toggleDebugSpawnZones(true)
                         this.gameWorld.toggleDebugBoundaries(true)
 
                         // Update UI to reflect changes
@@ -518,7 +507,6 @@ export class AppOne {
                         weaponRangeController.updateDisplay()
                         velocityVectorsController.updateDisplay()
                         restrictedZonesController.updateDisplay()
-                        spawnZonesController.updateDisplay()
                         boundariesController.updateDisplay()
                         console.log('🔍 All Debug Features Enabled')
                     },
@@ -537,7 +525,6 @@ export class AppOne {
                         weaponRangeState.enabled = false
                         velocityVectorsState.enabled = false
                         restrictedZonesState.enabled = false
-                        spawnZonesState.enabled = false
                         boundariesState.enabled = false
 
                         this.gameWorld.setDebugMode(false)
@@ -546,7 +533,6 @@ export class AppOne {
                         this.gameWorld.toggleDebugWeaponRange(false)
                         this.gameWorld.toggleDebugVelocityVectors(false)
                         this.gameWorld.toggleDebugRestrictedZones(false)
-                        this.gameWorld.toggleDebugSpawnZones(false)
                         this.gameWorld.toggleDebugBoundaries(false)
 
                         // Update UI to reflect changes
@@ -556,7 +542,6 @@ export class AppOne {
                         weaponRangeController.updateDisplay()
                         velocityVectorsController.updateDisplay()
                         restrictedZonesController.updateDisplay()
-                        spawnZonesController.updateDisplay()
                         boundariesController.updateDisplay()
                         console.log('🔍 All Debug Features Disabled')
                     },
