+++ src/ecs/Component.ts
@@ -326,7 +326,6 @@ export interface DebugComponent extends Component {
     showWeaponRange: boolean
     showVelocityVectors: boolean
     showRestrictedZones: boolean
-    showSpawnZones: boolean
     showBoundaries: boolean
 }
