+++ src/entities/DebugFactory.ts
@@ -14,7 +14,6 @@ export function createDebugEntity(world: World): Entity {
         showWeaponRange: false,
         showVelocityVectors: false,
         showRestrictedZones: false,
-        showSpawnZones: false,
         showBoundaries: false,
     }
