+++ src/systems/DebugSystem.ts
@@ -13,7 +13,6 @@ import {
 } from 'three'
 import { shipBoundaryConfig } from '../config/BoundaryConfig'
 import { restrictedZones } from '../config/RestrictedZoneConfig'
-import { getAllSpawnZones } from '../config/SpawnZoneConfig'
 import type {
     CollisionComponent,
     DebugComponent,
@@ -33,7 +32,6 @@ interface DebugGizmo {
         | 'weaponRange'
         | 'velocityVector'
         | 'restrictedZone'
-        | 'spawnZone'
         | 'boundary'
 }
 
@@ -45,7 +43,6 @@ export class DebugSystem extends System {
     private weaponRangeMaterial: MeshBasicMaterial
     private velocityMaterial: LineBasicMaterial
     private restrictedZoneMaterial: MeshBasicMaterial
-    private spawnZoneMaterial: MeshBasicMaterial
     private boundaryMaterial: MeshBasicMaterial
 
     constructor(world: World, scene: Scene) {
@@ -85,13 +82,6 @@ export class DebugSystem extends System {
             wireframe: true,
             transparent: true,
             opacity: 0.4,
-        })
-
-        this.spawnZoneMaterial = new MeshBasicMaterial({
-            color: 0x44ff44, // Green for spawn zones
-            wireframe: true,
-            transparent: true,
-            opacity: 0.3,
         })
 
         this.boundaryMaterial = new MeshBasicMaterial({
@@ -167,11 +157,6 @@ export class DebugSystem extends System {
         // Render restricted zones (only once, not per entity)
         if (debugComponent.showRestrictedZones) {
             this.renderRestrictedZones()
-        }
-
-        // Render spawn zones (only once, not per entity)
-        if (debugComponent.showSpawnZones) {
-            this.renderSpawnZones()
         }
 
         // Render boundaries (only once, not per entity)
@@ -459,17 +444,6 @@ export class DebugSystem extends System {
         }
     }
 
-    public toggleSpawnZones(enabled: boolean): void {
-        const debugEntities = this.getEntities()
-        if (debugEntities.length > 0) {
-            const debugComponent =
-                debugEntities[0].getComponent<DebugComponent>('debug')
-            if (debugComponent) {
-                debugComponent.showSpawnZones = enabled
-            }
-        }
-    }
-
     public toggleBoundaries(enabled: boolean): void {
         const debugEntities = this.getEntities()
         if (debugEntities.length > 0) {
@@ -515,40 +489,6 @@ export class DebugSystem extends System {
         }
     }
 
-    private renderSpawnZones(): void {
-        const spawnZones = getAllSpawnZones()
-        for (let i = 0; i < spawnZones.length; i++) {
-            const zone = spawnZones[i]
-
-            // Calculate zone dimensions
-            const width = zone.maxX - zone.minX
-            const depth = zone.maxZ - zone.minZ
-            const height = 1.5 // Slightly smaller height than restricted zones
-
-            // Calculate zone center
-            const centerX = (zone.minX + zone.maxX) / 2
-            const centerZ = (zone.minZ + zone.maxZ) / 2
-            const centerY = 0.5
-
-            // Create box geometry for the spawn zone
-            const geometry = new BoxGeometry(width, height, depth)
-            const mesh = new Mesh(geometry, this.spawnZoneMaterial)
-
-            // Position the mesh at the zone center
-            mesh.position.set(centerX, centerY, centerZ)
-
-            // Add to scene
-            this.scene.add(mesh)
-
-            // Track this gizmo for cleanup (use different negative ID range for spawn zones)
-            this.debugGizmos.push({
-                mesh,
-                entityId: -1000 - i - 1, // Different negative ID range for spawn zones
-                type: 'spawnZone',
-            })
-        }
-    }
-
     private renderBoundaries(): void {
         const boundary = shipBoundaryConfig
 
@@ -588,7 +528,6 @@ export class DebugSystem extends System {
         this.weaponRangeMaterial.dispose()
         this.velocityMaterial.dispose()
         this.restrictedZoneMaterial.dispose()
-        this.spawnZoneMaterial.dispose()
         this.boundaryMaterial.dispose()
     }
 }
